+++ tests/Header/CookieTest.php
@@ -6,8 +6,7 @@ namespace Tests\Innmind\Http\Header;
 use Innmind\Http\{
     Header\Cookie,
     Header,
-    Header\CookieValue,
-    Header\Parameter\Parameter
+    Header\Parameter\Parameter,
 };
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
@@ -15,19 +14,12 @@ class CookieTest extends TestCase
 {
     public function testInterface()
     {
-        $cookie = new Cookie(
-            $value = new CookieValue(new Parameter('foo', 'bar')),
+        $cookie = Cookie::of(
+            new Parameter('foo', 'bar'),
         );
 
-        $this->assertInstanceOf(Header\Custom::class, $cookie);
-        $this->assertSame('Cookie: foo=bar', $cookie->normalize()->toString());
-    }
-
-    public function testOf()
-    {
-        $cookie = Cookie::of(new Parameter('foo', 'bar'));
-
         $this->assertInstanceOf(Cookie::class, $cookie);
+        $this->assertInstanceOf(Header\Custom::class, $cookie);
         $this->assertSame('Cookie: foo=bar', $cookie->normalize()->toString());
     }
 }